State pyds9 is Python 3.5 or newer only
Use the python_requires argument for setup to state that this
module is for Python 3.5 or higher (but not version 4). I am not
sure what the minimum supported version should be, but Python 3.5
is coming to the end of its support period, so it seems to be a
useful cut off).