Tests: check for expected failure and re-enable skipped tests
Tests that check a failure case now explicitly check the
error that is raised and (generally) the error message.

This removes the "There were n expected failures" from the test
output.

A number of tests were marked as "skip", and its unclear to me why,
so let's run them and see what happens. It wasn't clear what the
actual test should be in one case, so I have tweaked it slightly
(test_ds9_openlist).