Travis: explicitly start xvfb
It appears that the default Xvfb setting on linux cause
tests to hang, so start it explicitly and with a screen
size and depth (matching that used for macOS).

It was found that Python 3.5 + AstroPy 3 (probably for
any AstroPy > 2) would hang, and so this combination has
been commented out.

It may be that SETUP_XVFB could be used here, but initial
testing was not successful, so this has not been explored
further at this time.

Most of the Linux tests now use a recent Ubuntu distribution
(bionic) and so a much-newer gcc/toolchain than before.

The Travis yml file is a mess and should be re-factored, but
I don't really want to touch it as I now (fingers crossed)
have something working.