Be explicit about conversion to Python strings (#64)
Fix up the conversion-to-string handling in the XPA layer, where
it looks like it is appropriate. This is based on #64

Two tests were added to ensure that users see an error message.
Before this PR the calls would error out in the XPA code when
trying to convert the error buffers to strings.