Try to reduce false warnings about ds9 already running
The aim is to only report if there's a ds9 process running, not
just something (such as xpans) running in a directory containing
the name ds9. This would be much better if we were to use a
structured representation of processess (e.g. as provided with
the psutil module), is currently UNIX specific, and is not
foolproof.